Add files via upload
The 0100 commands show the bit encoded information about what sensors are supported by car.
More detailed information on this webpage:
https://en.wikipedia.org/wiki/OBD-II_PIDs#Service_01_PID_00

Based on this, my suzuki swift hatchback answered this:
BE 3F B8 11

BE-
01 03 04 
05 06 07

3F-
0C 0B
0D 0E 0F 10

B8-
14 13 11
15

11-
1C
20

which mean it supports throttle. So now if the car is moving and I trample the accelerator pedal then it will calculate the consumption.
The long term consumption is stored in EEPROM. The long term consumption is calculated from 5 elements array. Every fifth cycle (which is a full cycle in short consumption array) average will show as an element in long term consumption. 

In keyboard handling, there is a little cheat for debugging purpose. If set the fuel to 1L and press the button five, it will reset the long term consumption array. Important, then set the correct fuel quantity to save the right fuel and nulled long term consumption array to EEPROM.